requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
biopython
requests
xlsxwriter
scipy

